Fixed the menu overflow issue
The menu was overflowing past the right hand side of the window creating
scrollable space that could be brought into view by certain iframed
widgets.

This fixes the bug by creating a container that prevents overflow.  It
has to use flexbox so the menu itself is 100% of the container height
and a scrollbar can be shown correctly.