RSS parsing bugfixes and enhancements
Parsing previously removed the contents of disallowed elements if they
were nested more than 2 deep, this fixes that by using jQuery's
`.unwrap()` method instead of `innerHTML`.

This commit also adds support for removing the first element in the
description if it's text content is identical to the item title and
slightly improves general parsing performance.