Made iChrome more cross-browser friendly
This is the result of a week long project to create a Firefox version.
Firefox's new WebExtensions API doesn't offer all the necessary APIs, so
it had to be built using the Addon SDK.  About half the necessary APIs
are built in an offline branch, but this should not continue to be
prioritized over other features.

This commit makes a number of changes, namely creating a central Browser
module that swaps in the correct API for the current browser, making
URLs relative in most cases, various CSS fixes, and the addition of
environment-specific widgets (such as Apps).  All global usage of
chrome.* APIs and localStorage is deprecated, except in the background
page which is Chrome-specific.