Redesigned the widget system
Although this update doesn't have a significant effect on performance,
it makes things more modular and easier to expand.  The system also has
better error handling, settings persistence, OAuth, and browser
permission handling.

It allows for easier development and adds support for minimized and
maximized widget views, along with isolated widget stylesheets and
resources. It also allows for things such as lazy-loaded widgets (not
rendered immediately if out of view) and widgets packaged into a single
JS file.

This commit also splits the build file since it was getting large.