Widget CSS: Switched from innerText to textContent
This change results in a 100x performance improvement (1.2s down to
12ms) in CSS processing (which only happens in the app during
development). I'm not sure why innerText has such bad performance in
this case, but given the performance it might be reparsing every rule in
the sheet after each rule is parsed (making things nearly O(n^2) for
each rule).