Auth: Fixed a bug where requests hitting invalid tokens wouldn't be issued correctly
When requests hit an invalid token a token refreshing request is issued.
By updating the request URL before this happens the conditional block
that adds the authorization header won't get triggered the second time
around.

In the case of the sync API this meant that two requests in a row would
fail (the first hitting a bad token and the second getting sent without
one at all) which used to sign the user out.