Widgets: Fixed multiple widget addition issues
When adding more than one of the same widget to a single column, the
previous system would move extra widgets to the end of the column and
then discard them on reload. This was happening because the widget
directory was using regular Backbone models instead of the proper widget
models that use the cid as their ID attribute.

This commit centralizes the modified Backbone models by creating a new
Registry.InstanceModel property.